api/requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
boto3==1.34.162
pyjwt==2.9.0
python-dotenv==1.0.1
